pyproject.toml
[tool.poetry]
name = "referral-service"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
redis = "^5.2.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
